test-v2/ai_module/requirements.txt
torch
numpy
pandas
scikit-learn
kafka-python
paramiko